pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["RayanMadan <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
PyNaCl = "^1.4.0"
Flask = "^2.0.1"
jokeapi = "^0.2.11"
dadjokes = "^1.3.2"
spotipy = "^2.18.0"
asyncpraw = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
